chore: resolve codacy
